test(core): colocate tests and migrate to test-utils
