+++ packages/core/src/prompts/confirm.test.ts
@@ -1,38 +1,31 @@
 import { cursor } from 'sisteransi';
-import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as ConfirmPrompt } from '../../src/prompts/confirm.js';
-import { MockReadable } from '../mock-readable.js';
-import { MockWritable } from '../mock-writable.js';
+import { beforeEach, describe, expect, test } from 'vitest';
+import { default as ConfirmPrompt } from './confirm.js';
+import { createMocks, type Mocks } from '@bomb.sh/tools/test-utils';
 
 describe('ConfirmPrompt', () => {
-	let input: MockReadable;
-	let output: MockWritable;
+	let mocks: Mocks<{ input: true; output: true }>;
 
 	beforeEach(() => {
-		input = new MockReadable();
-		output = new MockWritable();
-	});
-
-	afterEach(() => {
-		vi.restoreAllMocks();
+		mocks = createMocks({ input: true, output: true });
 	});
 
 	test('renders render() result', () => {
 		const instance = new ConfirmPrompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			active: 'yes',
 			inactive: 'no',
 		});
 		instance.prompt();
-		expect(output.buffer).to.deep.equal([cursor.hide, 'foo']);
+		expect(mocks.output.buffer).to.deep.equal([cursor.hide, 'foo']);
 	});
 
 	test('sets value and submits on confirm (y)', () => {
 		const instance = new ConfirmPrompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			active: 'yes',
 			inactive: 'no',
@@ -40,7 +33,7 @@ describe('ConfirmPrompt', () => {
 		});
 
 		instance.prompt();
-		input.emit('keypress', 'y', { name: 'y' });
+		mocks.input.emit('keypress', 'y', { name: 'y' });
 
 		expect(instance.value).to.equal(true);
 		expect(instance.state).to.equal('submit');
@@ -48,8 +41,8 @@ describe('ConfirmPrompt', () => {
 
 	test('sets value and submits on confirm (n)', () => {
 		const instance = new ConfirmPrompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			active: 'yes',
 			inactive: 'no',
@@ -57,7 +50,7 @@ describe('ConfirmPrompt', () => {
 		});
 
 		instance.prompt();
-		input.emit('keypress', 'n', { name: 'n' });
+		mocks.input.emit('keypress', 'n', { name: 'n' });
 
 		expect(instance.value).to.equal(false);
 		expect(instance.state).to.equal('submit');
@@ -66,8 +59,8 @@ describe('ConfirmPrompt', () => {
 	describe('cursor', () => {
 		test('cursor is 1 when inactive', () => {
 			const instance = new ConfirmPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				active: 'yes',
 				inactive: 'no',
@@ -75,14 +68,14 @@ describe('ConfirmPrompt', () => {
 			});
 
 			instance.prompt();
-			input.emit('keypress', '', { name: 'return' });
+			mocks.input.emit('keypress', '', { name: 'return' });
 			expect(instance.cursor).to.equal(1);
 		});
 
 		test('cursor is 0 when active', () => {
 			const instance = new ConfirmPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				active: 'yes',
 				inactive: 'no',
@@ -90,7 +83,7 @@ describe('ConfirmPrompt', () => {
 			});
 
 			instance.prompt();
-			input.emit('keypress', '', { name: 'return' });
+			mocks.input.emit('keypress', '', { name: 'return' });
 			expect(instance.cursor).to.equal(0);
 		});
 	});
