+++ packages/core/src/prompts/prompt.test.ts
@@ -1,68 +1,61 @@
 import { cursor } from 'sisteransi';
-import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as Prompt } from '../../src/prompts/prompt.js';
-import { isCancel } from '../../src/utils/index.js';
-import { MockReadable } from '../mock-readable.js';
-import { MockWritable } from '../mock-writable.js';
+import { beforeEach, describe, expect, test, vi } from 'vitest';
+import { default as Prompt } from './prompt.js';
+import { isCancel } from '../utils/index.js';
+import { createMocks, type Mocks } from '@bomb.sh/tools/test-utils';
 
 describe('Prompt', () => {
-	let input: MockReadable;
-	let output: MockWritable;
+	let mocks: Mocks<{ input: true; output: true }>;
 
 	beforeEach(() => {
-		input = new MockReadable();
-		output = new MockWritable();
-	});
-
-	afterEach(() => {
-		vi.restoreAllMocks();
+		mocks = createMocks({ input: true, output: true });
 	});
 
 	test('renders render() result', () => {
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 		// leave the promise hanging since we don't want to submit in this test
 		instance.prompt();
-		expect(output.buffer).to.deep.equal([cursor.hide, 'foo']);
+		expect(mocks.output.buffer).to.deep.equal([cursor.hide, 'foo']);
 	});
 
 	test('submits on return key', async () => {
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 		const resultPromise = instance.prompt();
-		input.emit('keypress', '', { name: 'return' });
+		mocks.input.emit('keypress', '', { name: 'return' });
 		const result = await resultPromise;
 		expect(result).to.equal(undefined);
 		expect(isCancel(result)).to.equal(false);
 		expect(instance.state).to.equal('submit');
-		expect(output.buffer).to.deep.equal([cursor.hide, 'foo', '\n', cursor.show]);
+		expect(mocks.output.buffer).to.deep.equal([cursor.hide, 'foo', '\n', cursor.show]);
 	});
 
 	test('cancels on ctrl-c', async () => {
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 		const resultPromise = instance.prompt();
-		input.emit('keypress', '\x03', { name: 'c' });
+		mocks.input.emit('keypress', '\x03', { name: 'c' });
 		const result = await resultPromise;
 		expect(isCancel(result)).to.equal(true);
 		expect(instance.state).to.equal('cancel');
-		expect(output.buffer).to.deep.equal([cursor.hide, 'foo', '\n', cursor.show]);
+		expect(mocks.output.buffer).to.deep.equal([cursor.hide, 'foo', '\n', cursor.show]);
 	});
 
 	test('does not write initialValue to value', () => {
 		const eventSpy = vi.fn();
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			initialValue: 'bananas',
 		});
@@ -75,23 +68,23 @@ describe('Prompt', () => {
 	test('re-renders on resize', () => {
 		const renderFn = vi.fn().mockImplementation(() => 'foo');
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: renderFn,
 		});
 		instance.prompt();
 
 		expect(renderFn).toHaveBeenCalledTimes(1);
 
-		output.emit('resize');
+		mocks.output.emit('resize');
 
 		expect(renderFn).toHaveBeenCalledTimes(2);
 	});
 
 	test('state is active after first render', async () => {
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 
@@ -105,8 +98,8 @@ describe('Prompt', () => {
 	test('emits truthy confirm on y press', () => {
 		const eventFn = vi.fn();
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 
@@ -114,7 +107,7 @@ describe('Prompt', () => {
 
 		instance.prompt();
 
-		input.emit('keypress', 'y', { name: 'y' });
+		mocks.input.emit('keypress', 'y', { name: 'y' });
 
 		expect(eventFn).toBeCalledWith(true);
 	});
@@ -122,8 +115,8 @@ describe('Prompt', () => {
 	test('emits falsey confirm on n press', () => {
 		const eventFn = vi.fn();
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 
@@ -131,7 +124,7 @@ describe('Prompt', () => {
 
 		instance.prompt();
 
-		input.emit('keypress', 'n', { name: 'n' });
+		mocks.input.emit('keypress', 'n', { name: 'n' });
 
 		expect(eventFn).toBeCalledWith(false);
 	});
@@ -139,8 +132,8 @@ describe('Prompt', () => {
 	test('emits key event for unknown chars', () => {
 		const eventSpy = vi.fn();
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 
@@ -148,7 +141,7 @@ describe('Prompt', () => {
 
 		instance.prompt();
 
-		input.emit('keypress', 'z', { name: 'z' });
+		mocks.input.emit('keypress', 'z', { name: 'z' });
 
 		expect(eventSpy).toBeCalledWith('z', { name: 'z' });
 	});
@@ -157,8 +150,8 @@ describe('Prompt', () => {
 		const keys = ['up', 'down', 'left', 'right'];
 		const eventSpy = vi.fn();
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 
@@ -167,7 +160,7 @@ describe('Prompt', () => {
 		instance.prompt();
 
 		for (const key of keys) {
-			input.emit('keypress', key, { name: key });
+			mocks.input.emit('keypress', key, { name: key });
 			expect(eventSpy).toBeCalledWith(key);
 		}
 	});
@@ -182,8 +175,8 @@ describe('Prompt', () => {
 		const eventSpy = vi.fn();
 		const instance = new Prompt(
 			{
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 			},
 			false
@@ -194,7 +187,7 @@ describe('Prompt', () => {
 		instance.prompt();
 
 		for (const [alias, key] of keys) {
-			input.emit('keypress', alias, { name: alias });
+			mocks.input.emit('keypress', alias, { name: alias });
 			expect(eventSpy).toBeCalledWith(key);
 		}
 	});
@@ -203,8 +196,8 @@ describe('Prompt', () => {
 		const abortController = new AbortController();
 
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			signal: abortController.signal,
 		});
@@ -223,8 +216,8 @@ describe('Prompt', () => {
 		abortController.abort();
 
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			signal: abortController.signal,
 		});
@@ -235,8 +228,8 @@ describe('Prompt', () => {
 
 	test('accepts invalid initial value', () => {
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			initialValue: 'invalid',
 			validate: (value) => (value === 'valid' ? undefined : 'must be valid'),
@@ -249,8 +242,8 @@ describe('Prompt', () => {
 
 	test('validates value on return', () => {
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			validate: (value) => (value === 'valid' ? undefined : 'must be valid'),
 		});
@@ -258,7 +251,7 @@ describe('Prompt', () => {
 
 		instance.value = 'invalid';
 
-		input.emit('keypress', '', { name: 'return' });
+		mocks.input.emit('keypress', '', { name: 'return' });
 
 		expect(instance.state).to.equal('error');
 		expect(instance.error).to.equal('must be valid');
@@ -266,15 +259,15 @@ describe('Prompt', () => {
 
 	test('validates value with Error object', () => {
 		const instance = new Prompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			validate: (value) => (value === 'valid' ? undefined : new Error('must be valid')),
 		});
 		instance.prompt();
 
 		instance.value = 'invalid';
-		input.emit('keypress', '', { name: 'return' });
+		mocks.input.emit('keypress', '', { name: 'return' });
 
 		expect(instance.state).to.equal('error');
 		expect(instance.error).to.equal('must be valid');
@@ -282,15 +275,15 @@ describe('Prompt', () => {
 
 	test('validates value with regex validation', () => {
 		const instance = new Prompt<string>({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			validate: (value) => (/^[A-Z]+$/.test(value ?? '') ? undefined : 'Invalid value'),
 		});
 		instance.prompt();
 
 		instance.value = 'Invalid Value $$$';
-		input.emit('keypress', '', { name: 'return' });
+		mocks.input.emit('keypress', '', { name: 'return' });
 
 		expect(instance.state).to.equal('error');
 		expect(instance.error).to.equal('Invalid value');
@@ -298,15 +291,15 @@ describe('Prompt', () => {
 
 	test('accepts valid value with regex validation', () => {
 		const instance = new Prompt<string>({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			validate: (value) => (/^[A-Z]+$/.test(value ?? '') ? undefined : 'Invalid value'),
 		});
 		instance.prompt();
 
 		instance.value = 'VALID';
-		input.emit('keypress', '', { name: 'return' });
+		mocks.input.emit('keypress', '', { name: 'return' });
 
 		expect(instance.state).to.equal('submit');
 		expect(instance.error).to.equal('');
