+++ packages/core/src/prompts/select.test.ts
@@ -1,38 +1,31 @@
 import { cursor } from 'sisteransi';
-import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as SelectPrompt } from '../../src/prompts/select.js';
-import { MockReadable } from '../mock-readable.js';
-import { MockWritable } from '../mock-writable.js';
+import { beforeEach, describe, expect, test } from 'vitest';
+import { default as SelectPrompt } from './select.js';
+import { createMocks, type Mocks } from '@bomb.sh/tools/test-utils';
 
 describe('SelectPrompt', () => {
-	let input: MockReadable;
-	let output: MockWritable;
+	let mocks: Mocks<{ input: true; output: true }>;
 
 	beforeEach(() => {
-		input = new MockReadable();
-		output = new MockWritable();
-	});
-
-	afterEach(() => {
-		vi.restoreAllMocks();
+		mocks = createMocks({ input: true, output: true });
 	});
 
 	test('renders render() result', () => {
 		const instance = new SelectPrompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			options: [{ value: 'foo' }, { value: 'bar' }],
 		});
 		instance.prompt();
-		expect(output.buffer).to.deep.equal([cursor.hide, 'foo']);
+		expect(mocks.output.buffer).to.deep.equal([cursor.hide, 'foo']);
 	});
 
 	describe('cursor', () => {
 		test('cursor is index of selected item', () => {
 			const instance = new SelectPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				options: [{ value: 'foo' }, { value: 'bar' }],
 			});
@@ -40,14 +33,14 @@ describe('SelectPrompt', () => {
 			instance.prompt();
 
 			expect(instance.cursor).to.equal(0);
-			input.emit('keypress', 'down', { name: 'down' });
+			mocks.input.emit('keypress', 'down', { name: 'down' });
 			expect(instance.cursor).to.equal(1);
 		});
 
 		test('cursor loops around', () => {
 			const instance = new SelectPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				options: [{ value: 'foo' }, { value: 'bar' }, { value: 'baz' }],
 			});
@@ -55,44 +48,44 @@ describe('SelectPrompt', () => {
 			instance.prompt();
 
 			expect(instance.cursor).to.equal(0);
-			input.emit('keypress', 'up', { name: 'up' });
+			mocks.input.emit('keypress', 'up', { name: 'up' });
 			expect(instance.cursor).to.equal(2);
-			input.emit('keypress', 'down', { name: 'down' });
+			mocks.input.emit('keypress', 'down', { name: 'down' });
 			expect(instance.cursor).to.equal(0);
 		});
 
 		test('left behaves as up', () => {
 			const instance = new SelectPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				options: [{ value: 'foo' }, { value: 'bar' }, { value: 'baz' }],
 			});
 
 			instance.prompt();
 
-			input.emit('keypress', 'left', { name: 'left' });
+			mocks.input.emit('keypress', 'left', { name: 'left' });
 			expect(instance.cursor).to.equal(2);
 		});
 
 		test('right behaves as down', () => {
 			const instance = new SelectPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				options: [{ value: 'foo' }, { value: 'bar' }],
 			});
 
 			instance.prompt();
 
-			input.emit('keypress', 'left', { name: 'left' });
+			mocks.input.emit('keypress', 'left', { name: 'left' });
 			expect(instance.cursor).to.equal(1);
 		});
 
 		test('initial value is selected', () => {
 			const instance = new SelectPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				options: [{ value: 'foo' }, { value: 'bar' }],
 				initialValue: 'bar',
@@ -103,35 +96,35 @@ describe('SelectPrompt', () => {
 
 		test('cursor skips disabled options (down)', () => {
 			const instance = new SelectPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				options: [{ value: 'foo' }, { value: 'bar', disabled: true }, { value: 'baz' }],
 			});
 			instance.prompt();
 			expect(instance.cursor).to.equal(0);
-			input.emit('keypress', 'down', { name: 'down' });
+			mocks.input.emit('keypress', 'down', { name: 'down' });
 			expect(instance.cursor).to.equal(2);
 		});
 
 		test('cursor skips disabled options (up)', () => {
 			const instance = new SelectPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				initialValue: 'baz',
 				options: [{ value: 'foo' }, { value: 'bar', disabled: true }, { value: 'baz' }],
 			});
 			instance.prompt();
 			expect(instance.cursor).to.equal(2);
-			input.emit('keypress', 'up', { name: 'up' });
+			mocks.input.emit('keypress', 'up', { name: 'up' });
 			expect(instance.cursor).to.equal(0);
 		});
 
 		test('cursor skips initial disabled option', () => {
 			const instance = new SelectPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				options: [{ value: 'foo', disabled: true }, { value: 'bar' }, { value: 'baz' }],
 			});
