+++ packages/core/src/prompts/text.test.ts
@@ -1,57 +1,50 @@
 import { styleText } from 'node:util';
 import { cursor } from 'sisteransi';
-import { afterEach, beforeEach, describe, expect, test, vi } from 'vitest';
-import { default as TextPrompt } from '../../src/prompts/text.js';
-import { MockReadable } from '../mock-readable.js';
-import { MockWritable } from '../mock-writable.js';
+import { beforeEach, describe, expect, test } from 'vitest';
+import { default as TextPrompt } from './text.js';
+import { createMocks, type Mocks } from '@bomb.sh/tools/test-utils';
 
 describe('TextPrompt', () => {
-	let input: MockReadable;
-	let output: MockWritable;
+	let mocks: Mocks<{ input: true; output: true }>;
 
 	beforeEach(() => {
-		input = new MockReadable();
-		output = new MockWritable();
-	});
-
-	afterEach(() => {
-		vi.restoreAllMocks();
+		mocks = createMocks({ input: true, output: true });
 	});
 
 	test('renders render() result', () => {
 		const instance = new TextPrompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 		});
 		// leave the promise hanging since we don't want to submit in this test
 		instance.prompt();
-		expect(output.buffer).to.deep.equal([cursor.hide, 'foo']);
+		expect(mocks.output.buffer).to.deep.equal([cursor.hide, 'foo']);
 	});
 
 	test('sets default value on finalize if no value', async () => {
 		const instance = new TextPrompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			defaultValue: 'bleep bloop',
 		});
 		const resultPromise = instance.prompt();
-		input.emit('keypress', '', { name: 'return' });
+		mocks.input.emit('keypress', '', { name: 'return' });
 		const result = await resultPromise;
 		expect(result).to.equal('bleep bloop');
 	});
 
 	test('keeps value on finalize', async () => {
 		const instance = new TextPrompt({
-			input,
-			output,
+			input: mocks.input,
+			output: mocks.output,
 			render: () => 'foo',
 			defaultValue: 'bleep bloop',
 		});
 		const resultPromise = instance.prompt();
-		input.emit('keypress', 'x', { name: 'x' });
-		input.emit('keypress', '', { name: 'return' });
+		mocks.input.emit('keypress', 'x', { name: 'x' });
+		mocks.input.emit('keypress', '', { name: 'return' });
 		const result = await resultPromise;
 		expect(result).to.equal('x');
 	});
@@ -59,8 +52,8 @@ describe('TextPrompt', () => {
 	describe('cursor', () => {
 		test('can get cursor', () => {
 			const instance = new TextPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 			});
 
@@ -71,69 +64,69 @@ describe('TextPrompt', () => {
 	describe('userInputWithCursor', () => {
 		test('returns value on submit', () => {
 			const instance = new TextPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 			});
 			instance.prompt();
-			input.emit('keypress', 'x', { name: 'x' });
-			input.emit('keypress', '', { name: 'return' });
+			mocks.input.emit('keypress', 'x', { name: 'x' });
+			mocks.input.emit('keypress', '', { name: 'return' });
 			expect(instance.userInputWithCursor).to.equal('x');
 		});
 
 		test('highlights cursor position', () => {
 			const instance = new TextPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 			});
 			instance.prompt();
 			const keys = 'foo';
 			for (let i = 0; i < keys.length; i++) {
-				input.emit('keypress', keys[i], { name: keys[i] });
+				mocks.input.emit('keypress', keys[i], { name: keys[i] });
 			}
-			input.emit('keypress', undefined, { name: 'left' });
+			mocks.input.emit('keypress', undefined, { name: 'left' });
 			expect(instance.userInputWithCursor).to.equal(`fo${styleText('inverse', 'o')}`);
 		});
 
 		test('shows cursor at end if beyond value', () => {
 			const instance = new TextPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 			});
 			instance.prompt();
 			const keys = 'foo';
 			for (let i = 0; i < keys.length; i++) {
-				input.emit('keypress', keys[i], { name: keys[i] });
+				mocks.input.emit('keypress', keys[i], { name: keys[i] });
 			}
-			input.emit('keypress', undefined, { name: 'right' });
+			mocks.input.emit('keypress', undefined, { name: 'right' });
 			expect(instance.userInputWithCursor).to.equal('foo█');
 		});
 
 		test('does not use placeholder as value when pressing enter', async () => {
 			const instance = new TextPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				placeholder: '  (hit Enter to use default)',
 				defaultValue: 'default-value',
 			});
 			const resultPromise = instance.prompt();
-			input.emit('keypress', '', { name: 'return' });
+			mocks.input.emit('keypress', '', { name: 'return' });
 			const result = await resultPromise;
 			expect(result).to.equal('default-value');
 		});
 
 		test('returns empty string when no value and no default', async () => {
 			const instance = new TextPrompt({
-				input,
-				output,
+				input: mocks.input,
+				output: mocks.output,
 				render: () => 'foo',
 				placeholder: '  (hit Enter to use default)',
 			});
 			const resultPromise = instance.prompt();
-			input.emit('keypress', '', { name: 'return' });
+			mocks.input.emit('keypress', '', { name: 'return' });
 			const result = await resultPromise;
 			expect(result).to.equal('');
 		});
