+++ packages/core/src/utils/cursor.test.ts
